Tree flattening
Adds Intermediate.java and Placeholder.java

The goal of Intermediate is to eventually generate the IR. Currently, it flattens expressions in the tree generated by Parser into placeholders. 

Placeholder is a class which helps transform complex expressions into three address instruction chains.

For an example of the flattened tree and what the placeholders hold, see flattenedTree in SampleOutputs