Statement List IR generation
Added in generation of IR for Statement lists.
These are composed of: 
	LabeledStatement
  	ExpressionStatement
  	SelectionStatement
  	IterationStatement
  	JumpStatement
Most, if not all, language features should now be handled by the IR.

Additionally, Compiler.java was modified to allow for the new -ir option which prints the IR to std out.

Several testing files were also updated or added.