Added Backend.java
This class is to serve as the converter between IR and assembly. While not fully implemented, this gets the core idea in motion.

The core idea of this converter is to switch between states as it encounters lines in the IR which invoke behaviors such as control flow. 

As we run linearly through the IR and find different behaviors we call a function to apply that behavior.

Additionally, this commit contains updates to Compiler and several test files for testing.